pyproject.toml
[tool.poetry]
name = "company-size-linkedin-scraper"
version = "0.1.0"
description = ""
authors = ["Paul Zhang <paulz@ucla.edu>"]
license = "GPL-3.0"
readme = "README.md"
packages = [{include = "company_size_linkedin_scraper"}]

[tool.poetry.dependencies]
python = "^3.11"
rich = "^13.4.2"
linkedin-api = {git = "https://github.com/tomquirk/linkedin-api.git"}
python-dotenv = "^1.0.0"
mypy = "^1.4.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.0.278"
black = "^23.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
